# otp_auth/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OTP Auth API",
    "version": "1.0.0",
    "description": "API for OTP-based authentication"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/auth/send-otp": {
      "post": {
        "summary": "Send OTP to phone",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string", "example": "+1234567890" }
                },
                "required": ["phone"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP Sent",
            "content": {
              "application/json": {
                "example": {
                  "message": "OTP sent successfully",
                  "phone": "+1234567890"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP and get JWT token",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string", "example": "+1234567890" },
                  "otp": { "type": "string", "example": "123456" }
                },
                "required": ["phone", "otp"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.xxxxx"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP"
          }
        }
      }
    },
    "/profile": {
      "get": {
        "summary": "Access protected route",
        "tags": ["Authentication"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Protected data",
            "content": {
              "application/json": {
                "example": {
                  "message": "Access granted to protected route",
                  "userId": "1"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}